move IntegrationWebhook component to app
